Refactor (KJH) : PostService생성, postupdate합치기, 코드 간소화 #1
