+++ src/main/java/com/example/meongnyangbook/S3/service/S3Service.java
@@ -1,4 +1,4 @@
-package com.example.meongnyangbook.S3;
+package com.example.meongnyangbook.S3.service;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
