+++ src/main/java/com/example/meongnyangbook/post/service/PostService.java
@@ -4,6 +4,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 public interface PostService {
 
-  public void update(Long id,
+  void update(Long id,
       MultipartFile[] multipartFiles, String[] deleteFileNames);
 }
